+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix Safari channel controller bug.
 - Fix Safari heatmap display bug.
 - Heatmap color canvas has precedence over selection canvas when resizing.
+- Updated the `cell-sets.json` schemas to allow both leaf nodes and non-leaf nodes in the same tree level.
 
 ## [0.1.6](https://www.npmjs.com/package/vitessce/v/0.1.6) - 2020-06-23