+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 
 ## Added
 - Testing protocol calls for all three browsers now.
-- Trevor's slides from NLM conference to README.md
+- Added a `cell-sets.json` schema version `0.1.3` to support probabilistic cell set assignments, with backwards compatibility with schema version `0.1.2`.
+- Trevor's slides from NLM conference to README.md.
 
 ### Changed
 - Fix Safari channel controller bug.
 - Fix Safari heatmap display bug.
 - Heatmap color canvas has precedence over selection canvas when resizing.
+- Updated the `cell-sets.json` schemas to allow both leaf nodes and non-leaf nodes in the same tree level.
 
 ## [0.1.6](https://www.npmjs.com/package/vitessce/v/0.1.6) - 2020-06-23